# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/projects":  {
                                        "post":  {
                                                     "operationId":  "StubsController_createProject",
                                                     "parameters":  [

                                                                    ],
                                                     "responses":  {
                                                                       "201":  {
                                                                                   "description":  ""
                                                                               }
                                                                   },
                                                     "tags":  [
                                                                  "Stubs"
                                                              ]
                                                 },
                                        "get":  {
                                                    "operationId":  "ProjectsController_findAll",
                                                    "parameters":  [
                                                                       {
                                                                           "name":  "clientId",
                                                                           "required":  true,
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "string"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "status",
                                                                           "required":  true,
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "string"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "page",
                                                                           "required":  true,
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "string"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "limit",
                                                                           "required":  true,
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "string"
                                                                                      }
                                                                       }
                                                                   ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  ""
                                                                              }
                                                                  },
                                                    "tags":  [
                                                                 "Projects"
                                                             ]
                                                }
                                    },
...
                  "/api/crm/leads/{id}/stage":  {
                                                    "patch":  {
                                                                  "operationId":  "StubsController_moveLeadStage",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "id",
                                                                                         "required":  true,
                                                                                         "in":  "path",
                                                                                         "schema":  {
                                                                                                        "type":  "string"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  ""
                                                                                            }
                                                                                },
                                                                  "tags":  [
                                                                               "Stubs"
                                                                           ]
                                                              }
                                                },
                  "/api/crm/leads/{id}/convert":  {
                                                      "post":  {
                                                                   "operationId":  "StubsController_convertLead",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "201":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 },
                                                                   "tags":  [
                                                                                "Stubs"
                                                                            ]
                                                               }
                                                  },
...
                  "/api/billing/invoices/{id}/payments":  {
                                                              "post":  {
                                                                           "operationId":  "StubsController_recordPayment",
                                                                           "parameters":  [
                                                                                              {
                                                                                                  "name":  "id",
                                                                                                  "required":  true,
                                                                                                  "in":  "path",
                                                                                                  "schema":  {
                                                                                                                 "type":  "string"
                                                                                                             }
                                                                                              }
                                                                                          ],
                                                                           "responses":  {
                                                                                             "201":  {
                                                                                                         "description":  ""
                                                                                                     }
                                                                                         },
                                                                           "tags":  [
                                                                                        "Stubs"
                                                                                    ]
                                                                       }
                                                          },
                  "/api/projects/{id}/tasks":  {
                                                   "post":  {
                                                                "operationId":  "StubsController_createTask",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "id",
                                                                                       "required":  true,
                                                                                       "in":  "path",
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "201":  {
                                                                                              "description":  ""
                                                                                          }
                                                                              },
                                                                "tags":  [
                                                                             "Stubs"
                                                                         ]
                                                            }
                                               },
                  "/api/tasks/{id}/move":  {
                                               "patch":  {
                                                             "operationId":  "StubsController_moveTask",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "required":  true,
                                                                                    "in":  "path",
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  ""
                                                                                       }
                                                                           },
                                                             "tags":  [
                                                                          "Stubs"
                                                                      ]
                                                         }
                                           },
...
              },
    "info":  {
                 "title":  "Trifusion-Dynamics Auth Service",
                 "description":  "The API description for Auth Service",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
...
                   }
}